move the matrix sorter code into a separate file
